+++ resources/views/welcome.blade.php
@@ -73,5 +73,6 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
     <script src="https://unpkg.com/leaflet@1.9.3/dist/leaflet.js"></script>
     <script src="{{ asset('js/mapa.js') }}"></script>
+    <script src="{{ asset('js/mapa_tecnico.js') }}"></script>
 </body>
 </html>
